requirements.txt
SpeechRecognition
pyttsx3
pyaudio
setuptools
text-to-speech
langchain_core
python-dotenv
langchain_community
sentence_transformers
langchain_huggingface
faiss-cpu
google-api-python-client